app/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-classic
chromadb
pypdf
openai-whisper
